requirements.txt
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
python-dotenv==1.0.0
matplotlib==3.8.2
seaborn==0.13.0
jupyter==1.0.0 